removes the testsuite submodule
The main reason is that opam recursively initializes submodules when
it installs bap from sources and given that our testsuite has a few
binaries along with the traces it sums up to 5Gb of space, which leads
to the out-of-disk-space errors when we build from docker. The easiest
solution would be to remove testsuite from the main repo. Moreover,
having it as a submodule doesn't really help and is mostly a
nuisance.

The proposed solution is to just clone the testsuite (potentially we
can select a branch or a commit) in the test scripts. This is
basically the same as what is git doing for submodules but now it will
be under our control.