disables the emission of section, file, and other symbols for ELF
The section symbols confuse our symbolization procedure so that at the
end we get symbols named as `.text` instead of their real names. To
prevent this, we are not emitting anymore relocation information about
symbols that are sections, files, objects, and other non-functions.

This commit fixes the nightly regression test. The underlying issue is
triggered only on modern LLVM (e.g., it is not observable on LLVM 6
and before, but is definitely triggered on 9, not sure about the in
between versions).