removes section and other symbols in ELF (#1220)
* removes an extra llvm from opam's depext

We already have llvm-9 installed so there is no reason to install the
default llvm also.

* enables options in configure-omake

though careful many of the options will be just ignored, either by the
script or by oasis/omake intergration.

Use the new feature immediately in the build-and-test action to enable
tests and specify the llvm version (which is for now ignored, but see
the next commit).

* pulls the llvm configuration from setup in omake

This is much better than specifying them manually as we did before :)

Also, switches back to the static linking with llvm when omake is
used.

* disables the emission of section, file, and other symbols for ELF

The section symbols confuse our symbolization procedure so that at the
end we get symbols named as `.text` instead of their real names. To
prevent this, we are not emitting anymore relocation information about
symbols that are sections, files, objects, and other non-functions.

This commit fixes the nightly regression test. The underlying issue is
triggered only on modern LLVM (e.g., it is not observable on LLVM 6
and before, but is definitely triggered on 9, not sure about the in
between versions).