switches to Fowler-Noll-Vo hash algorithm for hashing names (#1221)
The previous algorithm had a very bad collision rate, especially for
small strings. The new one is much better and is tested on large
dictionaries (of English words and password) and is guaranteed not
to collide on small strings.

Warning: the change of the hash function will break the knowledge base
format so do `bap --cache-clean` after the update.