optimizes the merge function for OGRE documents
The new implementation will untouch the document if the other one is
an empty document or is the same. When the documents are different and
non-trivial we will take the larger one and update it with the smaller
one, which is also faster. At the end, it makes merge more than 10
times faster under certain scenarios.