upgrades bap to LLVM 11 (#1229)
There were a few breaking changes so we have stick with if/else
macros. Mostly the the COFF loader is affected.

Note, I was using llvm-11 from the http://apt.llvm.org/bionic/ source
repository, which could be broken as it is missing libraries for
static linking, so in order to use this pre-built binaries you have to
use the dynamic linking mode, e.g.,
```
./configure-omake --with-llvm-version=11 \
                  --with-llvm-config=llvm-config-11
                  --disable-llvm-static
```

I hope this will go away with an official build of llvm.