integrates the thumb plugin with BAP build system
Right now we put thumb as a separate plugin, but we will probably move
it into the ARM plugin before the release, as there is no
reason (logical) to have it in a separate plugin. But lets move in
small steps.

It is recommended to use [omake][1] for rapid builds.

[1]: https://github.com/BinaryAnalysisPlatform/bap/wiki/Experimental-OMake-backend