removes exceptions and implements a proper plugin registration
Removes exceptions, we don't use them in general, only to indicate a
programmers error. Turned everything into failwith and we will,
eventually, get rid of all of them.

The lifter is now using encodings instead of architectures. That
ensures that it will be called in proper times. Also, makes the loader
invocation a little bit more straightforward.

Finally, implements a proper plugin extension registration - we shall
not register the lifter until we get the permission.