enables predicates when the disassembler is switched
When the disassembler is switched it may not have the predicates set
correctly. Ideally, we should track predicates per disassembler/per
state instead of resetting it on every switch, but benchmarking
doesn't show any significant performance impact, so we will just keep
it simple.

Also fixes a bug in the `with_preds` function, in which the condition
was flipped, i.e., predicates should be reset if there are predicates
to drop not the other way around.