uses MC insn in Thumb lifter; tweaks bil_lifter in the error case
Uses the machine code (MC) instruction information in the lifter
result and returns an empty instruction from the bil lifter when
no semantics is available to prevent conflicts in the properties.
(It used to work when the BIL lifter was the only one, but is no
longer a good idea).