removes emission of the non-functional symbols in the llvm backend
This is a temporary solution, later, when we will be ready for
non-functional symbols we will extend the scheme and add the symbol
type. That will enable us to ignore data symbols and do not treat them
as functions. In this case, constant islands were marked as arm code
and we were falling into them as thumb, which ended up in a conflict.