adds a proper handling of unitialized memory in Primus.Memory.map (#1230)
This PR fixes a bug in the Primus.Memory.map function that was
triggered when it is called before the current memory descriptor is
properly initialized. Instead of creating a new empty state it was
just silently ignoring the `map` operation. The bug was also affecting the
`Primus.Memory.del` operation, for which such behavior (i.e.,
ignoring) is probably fine. The new implementation fixes the update
function used by both, map and del to prevent any further misuses of
this function.