KB tweak: stops promises from trigger for non-empty properties
This tweak changes the semantics of the knowledge base by preventing
KB from triggering for properties that are already computed. In
certain cituations it may signficantly improve performance, i.e., when
the value of the property is already known and we do not want to
trigger expensive computations to recompute it.

P.S. to be honest, I was long time under an impression that this is
how KB works and was a bit surprised that it was only in my
imagination. We will use this feature to reduce the number of
unecessary calls for scoped objects (in fact it enables us to hoist
loop invariants from the loops).