optimizes the basic level of the disassembler
Several micro-optimizations. First, we will represent assembler
strings as bytes to minimize bytes to string transformations. Second,
we collect predicates using fold instead of filter map, that reduces
the number of allocations to 1/3 of what we had. Finally, we use the
new `Memory.view_exn` function to reduce the number of allocations
even further.