fixes a performance degradation
Finally found the culprit! The problem was that we were building each
graph for each subroutine from a clean slate, because we were
forgetting the previously built graphs.

To solve the problem we had to extend the interface of the explore
function so that it can explore from several starts while preserving
the state. Now bap is back on 2.1 speeds (maybe even faster).