updates the changes to the newest version of BAP
To address the ppx_js_style issue we have created a new package
ppx_bap that is analog to ppx_jane but with our own set of
rewriters (tha obviously excludes ppx_js_style as well as a bunch of
other rewriters that we don't want or need).

See https://github.com/BinaryAnalysisPlatform/ppx_bap for the package
details.

I have also updated the opam/opam file that is used to fetch bap
dependencies and added our opam-repo to the context of the
build-dev-repo action so that it can fetch the ppx_bap package that is
not yet in the upstream opam-repository (the PR will follow soon).