fixes warnings introduced with #1116 (#1236)
We missed some deprecation warnings in #1116. Most of them are dealing
with `sexp_*` types, which are easy to deal.

Another class of warnings is the deprecation of `Binprot.Of_binable`
due to introduction of uuid-identifiers witnessing the binary
representation. The idea is good and not novel to BAP, we always had
have those uuids. However, right now we're resolving this issue by
hushing the warning and we plan to address it in the nearest future,
in the next release.