initializes the default policy in primus taint analyzers (#1237)
Fixes #1144. We add three more components for policy selection. One
that selects the policy selected by the user via the command-line and
another two per each policy that we support. The base-taint-analyzer
system is leaving the policy selection unspecified. The taint-analyzer
system selects the policy based on the command-line parameter and teh
exact-taint-analyzer is relying on the exact taint propagation.