fixes improper compartmentalization of project computations
This PR fixes a bug reported by @jtpaash (with a nicely reproducible
demo). The bug was introduced when we switched to lazy computations so
that now symtab and program data structures computations were delayed
and could be run in a different state than the disassembly. The main
culprit was that those computations were forgetting to properly open
the current package to be able to access the objects of the currently
disassembled program (they were instead picking those objects from the
lastly disassembled program). Fortunately the fix is obvious and
simple.