makes the taint-attached observation on taint introductions (#1241)
The observation was missing when the taint was introduced with
new_direct or new_indirect operators. Although known for a long time,
this behavior was considered more as a feature rather than a bug. We
finally decided to rectify it and post the observation when a new
taint is introduced to make the behavior match with the semantics
described in the low-level terms. It also makes it much easier to
debug taint-related issues.