enables intermachine communication (#1243)
This PR adds a new interface, `Machine.Other`, that enables access to
the local state of other machines. It is possible both to pry into
others local state and even change it. So is it carefuly.

This PR is a step towards the solution of the #1076 issue. More will
come.