tweaks the Taint Engine and partially rewrites the Taint GC (#1244)
* enables intermachine communication

This PR adds a new interface, `Machine.Other`, that enables access to
the local state of other machines. It is possible both to pry into
others local state and even change it. So is it carefuly.

This PR is a step towards the solution of the #1076 issue. More will
come.

* tweaks the Taint Engine and partially rewrites the Taint GC

This PR resolves #1076 and enables better handling of the may
analysis (analysis that require analyzing all paths).

The Taint Engine now tracks taint objects globally across machines,
unlike previously when each machine had its own set of taint objects,
and taint object identifier was specific to a machine. This enables a
new implementation of the Taint Garbage Collection and refinement of
the taint-finalize observation. The GC is now folding over all
machines to check if the taint is unreachable in all machines. The
observation is triggered with live=true if the taint becomes
unreachable in the current machine but is still reachable in others,
and with live=false, when it comes out of existence in all machines,
or when the system terminates. This approach enables both may and must
analysis as we can both track liveness per machine and per
system (i.e., for all paths up to the linearity constraint).

The garbage collector is now run each machine instruction to prevent
undesired interference with other observations.

This PR also changes the semantics of the sanitize operation. Now,
when an object is sanitized, it is no longer tracked and removed from
the tracker (i.e., detached from all values and addresses). Before
that it was only detached from the selected value.