optimizes Primus memory operations
The memory operation is checking the layering information before any
read or write operation to ensure that the address being accessed is
properly mapped. It is slow and is linear to the number of layers.
However, if the address is already in the values mapping then it means
that it was checked and is writable and readable so we can omit the
layers check. The latter is only needed when the address is accessed
for the first time.