enables multiple stub regions in the stub resolver
So far we were assuming that there is only one section for stubs,
but in gcc FCF protection scheme they use `.plt.sec` in _addition_
to `.plt` to represent secured stubs. Therefore we need to keep track
of both, so we upgraded the representation from a simple interval to
an interval tree.

In addition, we added `.plt.sec` to the known names of stubs.