adds bap-references and uses them in plugins to access relocations
We were using the futures interface to acccess relocation information
stored in the ogre specification. That doesn't play nice with multiple
projects (well, at least in the way that it was implemented). Since the
references are already computed and stored in the knowledge base in
the relocatable plugin we can reuse them also in the BIL plugin. We
just need to make this library public.