fixes nightly tests (#1258)
The tests are requiring the presence of radare (for the sole purpose
of disabling radare), therefore they were failing since when we excluded
bap-radare2 from the default distribution of bap. The solution is
simple just install bap-radare. As a nice side-effect we will also
test that radare2 is properly installed.