fixes PE/COFF sections (#1263)
* fixes PE/COFF sections

We were missing non-executable sections in coff files. Thanks to
@TodAmon and @ReversingWithMe for noticing this.

* hushes non-executable sections for object files

The caveat with object files is that they have data (including .data,
.bss, .rdata, etc) sections all having the same zero address as the
.text section so we can't load them all together. Probably it is
expected that a loader will set the addresses but we need more
information on how to handle them correctly. For starters, we are just
disabling data sections in the object files.

* implements a temporary workaround to enable macOS builds

It looks like it takes a bit of time to fix the issue upstream. We
will remove those two lines as soon as
https://github.com/actions/virtual-environments/issues/2391 is
resolved.