introduces BIL special encodings and publishes BIL CT parser (#1265)
It is now possible to (ab)use BIL special statements to encode arbitrary
semantics. The payload is still stored as a string, but it is now
specially encoded and printed. Each kind of payload is represented by
an attribute, e.g., `Bil.call` is used to encode calls, and
`Bil.intrinsic` is used to encode intrinsics.

The Core Theory BIL parser, that enables reflection of the BIL
programs into the CT terms, is moved to the library and is made
available in `Bil.Theory` module. We might later extend this module
with the actual BIL theory, though it is already could be instantiated
via the `Theory.instance` function.