enables representation of an instruction as a BIL program
We used BIL as a language for specifying the semantics of an
instruction, but we may also (and now can) represent the instruction
itself using BIL.

The rules of the knowledge base are adopted and split accordingly, so
that the rule that uses the old lifters to produce the BIL code is
separated from the rule that computes the semantics of the BIL code.

It is now possible to get easily an `Insn.t` value from `bil` and even
to go from bil to sub or program term.

E.g., to get the semantics of BIL just create a new program object and
set its bil code,

```ocaml
let create_insn bil  =
  KB.Object.create Theory.Program.cls >>= fun obj ->
  KB.provide Bil.code obj bil >>| fun () ->
  obj
```

You can run this computation, for example, in the BAP toplevel
knowledge base, e.g.,

```ocaml
let insn_of_bil bil : insn =
  Toplevel.eval Theory.Semantics.slot @@ create_insn bil
```

To get IR from insn use the `Blk.from_insn` function, so the full path
from BIL to BIR will be

```
let bir_of_bil : blk term list =
  Blk.from_insn insn_of_bil @@ create_insn bil
```