drafts a new implementation of the lisp semantics
The new implementation is more tightly integrated with the knowledge
base and will totally rely on promises to provide semantics of lisp
definitions as well as lisp primitives. There will not be anymore the
special type for the semantics of a lisp program, instead a KB value
of Theory.Semantics.cls will be used. The class was extended with the
value slot, so no effects can have values.