optimizes primus lisp semantics computation
By removing a request to compute a name when the unit is not a lisp
unit. It helps since the name computation involves the stub resolver,
which in turn requires the semantics of the nearby instructions, and
so on.

In general, the rel-symbolizer shall be rewritten as it is so easy to
break it.