adds a Primus Lisp fronted  (#1268)
* implements some basic operations

* implements primus semantics primitives

* publishes the semantics module

primitives are still not published

* passes the target to the program and uses it in reification

* adds the top value and the experimental match module

* drafts a new implementation of the lisp semantics

The new implementation is more tightly integrated with the knowledge
base and will totally rely on promises to provide semantics of lisp
definitions as well as lisp primitives. There will not be anymore the
special type for the semantics of a lisp program, instead a KB value
of Theory.Semantics.cls will be used. The class was extended with the
value slot, so no effects can have values.

* drops the no longer necessary target from the lisp program

* implements equality on primus lisp programs and sources

right now the equality is purely textual equality.

* implements the semantics promise that provides reification of lisp

* publishes the minimal interface

* publishes a KB.Value.resort function

* prevents bir reifier from generating empty graphs.

* more fixes to the lisp lifter

* integrates semantic primitives with the main lisp module

and implements a few primitives, far from complete set though.

* adds more primitives

* we have an MVP Lisp compiler

* rectifies the interface a little bit, introduces a few helpers

* adds documentation and prettifies the interface a bit

* optimizes primus lisp semantics computation

By removing a request to compute a name when the unit is not a lisp
unit. It helps since the name computation involves the stub resolver,
which in turn requires the semantics of the nearby instructions, and
so on.

In general, the rel-symbolizer shall be rewritten as it is so easy to
break it.